fixed orders

+++ src/app/api/menu-items/route.js
@@ -4,14 +4,21 @@ import mongoose from "mongoose";
 export async function POST(req) {
   mongoose.connect(process.env.MONGO_URL);
   const data = await req.json();
-  const menuItemDoc = await MenuItems.create(data);
-  return Response.json(menuItemDoc);
+  if (await isAdmin()) {
+    const menuItemDoc = await MenuItems.create(data);
+    return Response.json(menuItemDoc);
+  } else {
+    return Response.json({});
+  }
 }
 
 export async function PUT(req) {
   mongoose.connect(process.env.MONGO_URL);
-  const { _id, ...data } = await req.json();
-  await MenuItems.findByIdAndUpdate(_id, data);
+  if (await isAdmin()) {
+    const { _id, ...data } = await req.json();
+    await MenuItems.findByIdAndUpdate(_id, data);
+  }
+
   return Response.json(true);
 }
 
@@ -19,7 +26,10 @@ export async function DELETE(req) {
   mongoose.connect(process.env.MONGO_URL);
   const url = new URL(req.url);
   const _id = url.searchParams.get("_id");
-  await MenuItems.deleteOne({ _id });
+  if (await isAdmin()) {
+    await MenuItems.deleteOne({ _id });
+  }
+
   return Response.json(true);
 }
